# backend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "GuildSettingsDto" : {
        "properties" : {
          "primaryLocationLat" : {
            "format" : "double",
            "type" : "number"
          },
          "primaryLocationLng" : {
            "format" : "double",
            "type" : "number"
          },
          "primaryLocationName" : {
            "type" : "string"
          },
          "primaryLocationPlaceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "GuildSettingsRequest" : {
        "properties" : {
          "primaryLocationLat" : {
            "format" : "double",
            "type" : "number"
          },
          "primaryLocationLng" : {
            "format" : "double",
            "type" : "number"
          },
          "primaryLocationName" : {
            "type" : "string"
          },
          "primaryLocationPlaceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/guild/{guildId}/settings" : {
      "get" : {
        "operationId" : "getSettings",
        "parameters" : [ {
          "in" : "path",
          "name" : "guildId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuildSettingsDto"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "tags" : [ "Guild" ]
      },
      "patch" : {
        "operationId" : "updateSettings",
        "parameters" : [ {
          "in" : "path",
          "name" : "guildId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GuildSettingsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuildSettingsDto"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "tags" : [ "Guild" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}